third_party/WebKit/LayoutTests/external/wpt/dom/events/Event-defaultPrevented-expected.txt

This is a testharness.js-based test.
PASS When an event is created, defaultPrevented should be initialized to false.
PASS initEvent should work correctly (not cancelable).
PASS preventDefault() should not change defaultPrevented if cancelable is false.
FAIL returnValue should not change defaultPrevented if cancelable is false. assert_equals: defaultPrevented expected false but got true
PASS initEvent should work correctly (cancelable).
PASS preventDefault() should change defaultPrevented if cancelable is true.
PASS returnValue should change defaultPrevented if cancelable is true.
PASS initEvent should unset defaultPrevented.
Harness: the test ran to completion.
